Inspections (#26)
* Add serial version to serialisable classes.  Then don't use it properly.

* Convert fields to local

* Make final-able things final

* Mis sorted modifiers

* Redundant field initialisation

* Nullsafe comparison ordering

* Remove unnecessary super()

* Unnecessarily qualified static import

* Configured control flow issues

* Inverted if statement

* Fix non final field in exception

* Ensure exception root causes are passed on

* Public abstract constructors become protected

* Javadoc line breaks

* Use pattern variables

* Use String.valueOf() rather than +""

* String.equals versus String.isEmpty

* Inherit variable names hierarchically

* Variable name fixup

* Numeric cleanups

* entrySet iteration

* indexOf char rather than String where possible

* String Builders

* serialisations

* Inspections

* Remove duplicates

* Make final

* Use isEmpty()

+++ src/net/coagulate/Core/Tools/UnixTime.java
@@ -244,7 +244,7 @@ public abstract class UnixTime {
 	@Nonnull
 	private static String fromUnixTime(final int date,
 	                                   @Nonnull final DateFormat df) {
-		return df.format(new Date(((long) (date))*((long) 1000)));
+		return df.format(new Date((date) *(1000L)));
 	}
 
 }
